requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2
celery
redis
pydantic
python-dotenv
fastapi-users
fastapi-cache
fastapi-cache2
email-validator-2.2.0
python-dotenv
psycopg2
asyncpg
flower
pytest
httpx
coverage
pytest-asyncio